server: setup connection to local database
Will be creating initial models locally before deploying to production.
